requirements.txt
yt_dlp=2023.3.4
ffmpeg-python=0.2.0
openai=0.27.7
google-api-python-client=2.88.0